% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triviawidget.R
\name{triviawidget}
\alias{triviawidget}
\title{triviawidget}
\usage{
triviawidget(questions)
}
\arguments{
\item{questions}{A list of questions. Every question must have the properties listed below.
\describe{
  \item{text}{The question text}
  \item{choices}{A list of strings with the possible answers}
  \item{answer}{A list with the 0-based index correct choice or choices}
}}
}
\description{
Render a trivia based on a list of questions
}
